+++ src/components/HistoryEventDetails/layouts/CollectibleTransactionEventDetails.js
@@ -27,7 +27,7 @@ import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
 
 // Components
 import { Row, Spacing } from 'components/modern/Layout';
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import Text from 'components/core/Text';
 import TransactionStatusIcon from 'components/modern/TransactionStatusIcon';
