+++ src/components/HistoryEventDetails/layouts/EnsNameEventDetails.js
@@ -27,7 +27,7 @@ import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
 
 // Components
 import { Spacing } from 'components/modern/Layout';
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
