+++ src/components/legacy/Button/Button.js
@@ -238,7 +238,7 @@ const getVariant = (props) => {
 /**
  * @deprecated This compontent is considered legacy and should not be used in new code
  *
- * Use: components/modern/Button instead
+ * Use: components/core/Button instead
  */
 
 class Button extends React.Component<CombinedProps, State> {
