+++ src/screens/AddCash/AddCash.js
@@ -37,7 +37,7 @@ import { isLogV2AppEvents } from 'utils/environment';
 
 // Components
 import { Container } from 'components/modern/Layout';
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import TextInput from 'components/legacy/TextInput';
 import HeaderBlock from 'components/HeaderBlock';
 import Text from 'components/core/Text';
