+++ src/screens/Assets/components/PillarPaySummary.js
@@ -25,7 +25,7 @@ import styled from 'styled-components/native';
 import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import FiatValueView from 'components/modern/FiatValueView';
 import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
