+++ src/screens/Exchange/Exchange/OfferCard.js
@@ -23,7 +23,7 @@ import styled from 'styled-components/native';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import Text from 'components/core/Text';
 import DynamicSizeImage from 'components/DynamicSizeImage';
