+++ src/screens/WalletConnect/CallRequest/SignatureRequestContent.js
@@ -23,7 +23,7 @@ import { useTranslation } from 'translations/translate';
 import styled from 'styled-components/native';
 
 // Components
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
