+++ src/screens/WalletConnect/CallRequest/UnsupportedRequestContent.js
@@ -23,7 +23,7 @@ import styled from 'styled-components/native';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
