+++ src/screens/WalletConnect/Requests/WalletConnectConnectorRequestModal.js
@@ -25,7 +25,7 @@ import { Keyboard } from 'react-native';
 
 // Components
 import BottomModal from 'components/modern/BottomModal';
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
