+++ src/screens/WalletMigrationArchanova/SelectAssets/WalletMigrationArchanovaSelectAssets.js
@@ -27,7 +27,7 @@ import { isEmpty } from 'lodash';
 
 // Components
 import { Container, Footer } from 'components/modern/Layout';
-import Button from 'components/modern/Button';
+import Button from 'components/core/Button';
 import HeaderBlock from 'components/HeaderBlock';
 
 // Constants
